revno: 16322 message:   Fixed nullpointer vulnerability in TrackedEntityFormDeletionHandler
committer: Lars Helge Overland <larshelge@gmail.com>
branch nick: dhis2
timestamp: Tue 2014-08-05 15:37:40 +0200

+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/trackedentity/TrackedEntityFormDeletionHandler.java
@@ -62,6 +62,10 @@ public class TrackedEntityFormDeletionHandler
     public void deleteProgram( Program program )
     {
         TrackedEntityForm form = formService.getFormsWithProgram( program );
-        formService.deleteTrackedEntityForm( form );
+        
+        if ( form != null )
+        {
+            formService.deleteTrackedEntityForm( form );
+        }
     }
 }
